% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression using Ordinary Least Squares (OLS)}
\description{
Linear Regression using Ordinary Least Squares (OLS)
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula object defining the regression model.}

\item{\code{data}}{A data frame containing the variables for the regression.}

\item{\code{X}}{Design matrix (model matrix).}

\item{\code{y}}{Response variable.}

\item{\code{beta_hat}}{Coefficients estimated by OLS.}

\item{\code{y_hat}}{Fitted values.}

\item{\code{residuals}}{Residuals (differences between observed and fitted values).}

\item{\code{sigma_hat}}{Residual standard error.}

\item{\code{t_values}}{T-statistics for the coefficients.}

\item{\code{p_values}}{P-values associated with each coefficient}
}}


